# documentazione e database/openapi.yaml
openapi: 3.0.3
info:
  title: Fooody REST API
  description: API RESTful per la gestione delle ordinazioni e delle consegne a domicilio dell'attività Fooody.
  version: 1.0.0
  contact:
    name: Maikol Gaparroni
    email: maikol.gasparroni@student.univaq.it
servers:
  - url: http://localhost:8080/Fooody/api
    description: Server locale (Apache Tomcat)
paths:
...
  /caratteristiche:
    get:
      summary: Elenco di tutte le caratteristiche
      tags: [Caratteristiche]
      responses:
        '200':
          description: Lista delle caratteristiche
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Caratteristica'
    post:
      security:
        - bearerAuth: []
      summary: Crea una nuova caratteristica (opzione menu)
      tags: [Caratteristiche]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Caratteristica'
      responses:
        '201':
          description: Caratteristica creata con successo.
        '403':
          description: Accesso negato.
        '500':
          description: Errore interno del server.
  /caratteristiche/{idCaratteristica}:
    put:
      security:
        - bearerAuth: []
      summary: Aggiorna una caratteristica esistente
      tags: [Caratteristiche]
      parameters:
        - in: path
          name: idCaratteristica
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Caratteristica'
      responses:
        '200':
          description: Caratteristica aggiornata con successo.
        '403':
          description: Accesso negato.
    delete:
      security:
        - bearerAuth: []
      summary: Elimina una caratteristica
      tags: [Caratteristiche]
      parameters:
        - in: path
          name: idCaratteristica
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Caratteristica eliminata con successo.
        '403':
          description: Accesso negato.
        '409':
          description: Conflitto - Caratteristica in uso.
  /caratteristiche/gruppi:
    get:
      summary: Elenco di tutti i gruppi di mutua esclusione
      tags: [Caratteristiche]
      responses:
        '200':
          description: Lista dei gruppi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GruppoMutuaEsclusione'
    post:
      security:
        - bearerAuth: []
      summary: Crea un nuovo gruppo di mutua esclusione
      tags: [Caratteristiche]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GruppoMutuaEsclusione'
      responses:
        '201':
          description: Gruppo creato con successo.
        '403':
          description: Accesso negato.
        '500':
          description: Errore interno del server.
  /caratteristiche/gruppi/{idGmc}:
    put:
      security:
        - bearerAuth: []
      summary: Aggiorna un gruppo esistente
      tags: [Caratteristiche]
      parameters:
        - in: path
          name: idGmc
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GruppoMutuaEsclusione'
      responses:
        '200':
          description: Gruppo aggiornato con successo.
        '403':
          description: Accesso negato.
    delete:
      security:
        - bearerAuth: []
      summary: Elimina un gruppo di mutua esclusione
      tags: [Caratteristiche]
      parameters:
        - in: path
          name: idGmc
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Gruppo eliminato con successo.
        '403':
          description: Accesso negato.
        '409':
          description: Conflitto - Gruppo in uso da caratteristiche.

  # ------------------
  # Ordini
  # ------------------
...
components:
  schemas:
...
    GruppoMutuaEsclusione:
      type: object
      properties:
        idGmc:
          type: integer
          example: 1
        nome:
          type: string
          example: "Scelta Cottura"
        descrizione:
          type: string
          example: "Seleziona come preferisci la carne"
    #################################à
    Caratteristica:
      type: object
      properties:
        idCaratteristica:
          type: integer
          example: 1
        nome:
          type: string
          example: "Piccante"
        differenzaPrezzo:
          type: number
          format: float
          example: -0.50
        isDefault:
          type: boolean
          example: false
        idGmc:
          type: integer
          example: 2
        gruppo:
          type: string
          description: Il nome del gruppo di caratteristiche in mutua esclusione se presente
          example: "Zucchero"
    #############################################
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Utilizza un token JWT per autenticarsi nelle richieste alle risorse protette. Il token deve essere incluso nell'header Authorization con il prefisso "Bearer".